fix: resolve fresh production-readiness security audit findings (#249 #250 #251 #252 #253)
Independent 2026-07-06 production-readiness security audit (round 2, post-48e6f20).
Three parallel fresh audits (security-researcher + 2x security-auditor) plus an
adversarial re-audit found 5 defects the prior GO verdict missed; all fixed with
load-bearing regression tests (each verified fails-before/passes-after).

- HEIF-ILOC-EXTENT (HIGH, #249, CWE-770/834): readIlocFullExtents and
  readIlocSimpleExtents ran the full attacker-controlled extent_count (uint16, up
  to 65535) while only the pre-alloc was capped; with all iloc field-size nibbles
  0 each iteration consumed 0 input bytes -> ~300000x memory amplification (~8-20KB
  file -> multi-GB OOM on Inject; CPU-DoS on Extract). Guard the append at
  maxIlocExtentsPerItem, cap the loop bound in the zero-field-size degenerate case,
  and add a maxIlocItems=4096 outer item-count cap (mirrors parseBigTIFFIFD0).
  Re-audit: worse-than-spec PoC sub-ms; 44/44 real HEIF fixtures unaffected.

- ORF-RW2-DOUBLEWRITE (HIGH, #250, CWE-664): writeTIFFORF/writeTIFFRW2 passed
  m.EXIF directly to the mutating relocator instead of cloneEXIF(m.EXIF); a second
  Write() on the same *Metadata silently corrupted image data (diverged at byte
  653530 on Olympus E410.orf). Siblings were fixed by #109; ORF/RW2 (#104) were
  missed. Pass cloneEXIF at both sites. Re-audit: 51/51 real ORF/RW2 fixtures
  double-write byte-identical.

- WRITE-READALL-OOM (HIGH, #251, CWE-770): all 6 TIFF-family write paths used a
  bare io.ReadAll(r) when rawEXIF==nil; #140 capped format/* but missed these root
  sites, so NewMetadata()+Write(largeUntrustedStream) OOM'd. Add readAllCapped
  (io.LimitReader + ErrFileTooLarge) and a root-package maxFileSize (256 MiB),
  mirroring the format-package convention. Re-audit: infinite reader bounded to
  ~185ms/0MB heap.

- EXIF-BO-002 (MEDIUM, #252, CWE-198): IFD.set inherited the entry byteorder flag
  from Entries[0] (LE for a freshly-created empty GPSIFD/ExifIFD), so on a
  big-endian source the numeric accessors (GPS/ExposureTime/FNumber/ISO/
  FocalLength/ImageSize) returned garbage after Set* and before Encode() (sibling
  of the #246 fix, which only corrected value-byte encoding). Thread an explicit
  bigEndian bool (from ifd0BigEndian()) through all 21 set() call sites. Re-audit:
  Encode() output SHA-256 byte-identical pre/post for both LE and BE (only the
  in-memory decode flag changed).

- TIFF-INT-TRUNC-32 (LOW, #253, CWE-681/190): findExifIFDOffset in tiff/orf/rw2/
  cr3 (and cr3 mergeCMT) checked int(ifd0Off)+2>len(data) then indexed
  data[ifd0Off:]; on 32-bit platforms int(ifd0Off) is negative for ifd0Off>=2^31,
  the guard passes, and the slice panics. Compare in uint64 before int()
  conversion, mirroring the #74/#45 idiom. Re-audit: GOARCH=386 cross-build+vet
  clean.

Validation (independent re-run at fixed tree): gofmt clean; go vet clean;
go build clean; go test -race ./... PASS (21 pkgs, 0 races); golangci-lint 0
issues; govulncheck clean (Go 1.26.4); fuzz FuzzHEIFExtract/Inject + FuzzParseEXIF
+ FuzzRead 40s each, 0 crashers. Adversarial re-audit: all 5 findings CLOSED, 0
regressions; hot-path (HEIF extent loop, exif IFD.set) allocation profiles
byte-identical. Verdict: GO for production.

+++ format/raw/cr3/cr3.go
@@ -158,7 +158,14 @@ func mergeCMT(cmt1, cmt2 []byte) []byte {
 	}
 	exifIFDOffset := getExifIFDOffset(cmt1)
 	// If the ExifIFD offset is within CMT1, no merge needed.
-	if exifIFDOffset == 0 || int(exifIFDOffset) < len(cmt1) {
+	//
+	// Security audit FIX 5 (CWE-681/190): compare in uint64, not int. On a
+	// 32-bit platform (GOARCH=386/arm), int(exifIFDOffset) for an offset
+	// >= 2^31 is negative, which would make this comparison true for ANY
+	// len(cmt1) — silently skipping the CMT1+CMT2 merge and leaving the
+	// ExifIFD pointer dangling past the end of cmt1. Mirrors the identical
+	// fix in findExifIFDOffset below.
+	if exifIFDOffset == 0 || uint64(exifIFDOffset) < uint64(len(cmt1)) {
 		return cmt1
 	}
 	// ExifIFD pointer extends into CMT2: concatenate.
@@ -172,10 +179,18 @@ func mergeCMT(cmt1, cmt2 []byte) []byte {
 // 0x8769 (ExifIFD) and returns its LONG value (the offset). Returns 0 if not
 // found or if buf is too short to parse.
 func findExifIFDOffset(buf []byte, ifd0Off uint32, order binary.ByteOrder) uint32 {
-	if int(ifd0Off)+2 > len(buf) {
+	// Security audit FIX 5 (CWE-681/190): compare in uint64, not int, before
+	// converting ifd0Off to int. On a 32-bit platform (GOARCH=386/arm),
+	// int(ifd0Off) for ifd0Off >= 2^31 is negative, which would let a bad
+	// offset pass an int-typed bound check and then panic on buf[pos:].
+	// Mirrors the #74 fix in format/detect.go's parseClassicTIFFIFD0 and the
+	// #45 fix in format/jpeg's parseIRBEntry.
+	if uint64(ifd0Off)+2 > uint64(len(buf)) {
 		return 0
 	}
 	count := order.Uint16(buf[ifd0Off:])
+	// ifd0Off ≤ uint64(len(buf))-2, and len(buf) is a valid int on this
+	// platform, so ifd0Off < len(buf) fits int safely here.
 	pos := int(ifd0Off) + 2
 	for range int(count) {
 		if pos+12 > len(buf) {
